Add process pool for performing synchronous cocaine service requests
As of tornado 4.x, cocaine-framework-python v0.11 synchronous requests
with .get() method perfromed from running worker throw exception
saying that "IOLoop is already running". Due to the fact that
we are currently using Flask we cannot switch to using
asynchronous API at the moment.
A separate process pool is dedicated to performing cocaine
requests synchronously. Each process in a pool run
ioloop only for request duration and hence does not lead
to raising exception.